feat: Add custom attribute uuid for binary data
Add experiment to insert 1M rows into a table that holds uuid stored as
binary (16). Once the script completes, we report statistical
information regarding insertion and select during the run. Storing uuid
as binary data is not only efficient in terms of storage but also the
speed is not compromised either.